rename action to sendSchemaToInspector
