+++ packages/destination-actions/src/destinations/avo/sendSchemaToInspector/__tests__/index.test.ts
