+++ packages/destination-actions/src/destinations/avo/sendSchemaToInspector/__tests__/snapshot.test.ts
