+++ packages/destination-actions/src/destinations/avo/sendSchemaToInspector/avo.ts
