+++ packages/destination-actions/src/destinations/avo/sendSchemaToInspector/generated-types.ts
